requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
defusedxml==0.7.1
distlib==0.3.8
Django==4.2.9
django-filter==23.5
djangorestframework==3.14.0
factory-boy==3.3.0
Faker==22.1.0
filelock==3.13.1
httpie==3.2.2
idna==3.6
importlib-metadata==7.0.1
Markdown==3.5.1
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
pipenv==2023.11.15
platformdirs==4.1.0
Pygments==2.17.2
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.1.0
virtualenv==20.25.0
zipp==3.17.0
